be/requirements.txt
fastapi[standard]==0.118.0
uvicorn[standard]==0.24.0
instructor
pydantic
python-dotenv
ruff
motor
beanie